refactor(presentation): split pose detection screen
